Bot-Lokalisierung: BPS-Sensor (BPS_AVAILABLE) und neues Verhalten bot_scan_beacons (BEHAVIOUR_SCAN_BEACONS_AVAILABLE)

+++ bot-logic/behaviour_catch_pillar.c
@@ -372,7 +372,7 @@ static uint8_t turn_cancel_check(void) {
 	case 0:
 		/* Check mit linkem Sensor */
 		if (sensDistL <= MAX_PILLAR_DISTANCE) {
-			headingL = heading;
+			headingL = (int16_t) heading;
 			side = 1;
 			return False;
 		}
@@ -381,7 +381,7 @@ static uint8_t turn_cancel_check(void) {
 		/* Check mit rechtem Sensor */
 		if (sensDistR <= MAX_PILLAR_DISTANCE) {
 			catch_pillar_state = state_after_cancel;
-			headingR = heading;
+			headingR = (int16_t) heading;
 			return True;
 		}
 		break;
@@ -431,7 +431,7 @@ void bot_catch_pillar_behaviour(Behaviour_t * data) {
 		if (dHead < 0) {
 			dHead += 360;
 		}
-		int16_t dist = (DISTSENSOR_POS_SW * 2.0f) / dHead * (180.0f / M_PI);
+		int16_t dist = (int16_t) ((DISTSENSOR_POS_SW * 2.0f) / dHead * (180.0f / M_PI));
 		/* Objektkoordis berechnen */
 		obj_pos = calc_point_in_distance(headingR, /*DISTSENSOR_POS_FW + */ dist,
 				-DISTSENSOR_POS_SW);
